fix some issue and added parsefloat

+++ js/app.js
@@ -49,7 +49,7 @@ const addToCart = (id, price) => {
 // get input value function
 const getInputValue = (id) => {
   const element = document.getElementById(id).innerText;
-  const converted = parseInt(element);
+  const converted = parseFloat(element);
   return converted;
 };
